// packages/protocol/typescript/generated/yiru/runtime/v1/git_common_pb.d.ts
export declare type GitChangeEntry = Message<"yiru.runtime.v1.GitChangeEntry"> & {
  /**
   * @generated from field: string path = 1;
   */
  path: string;

  /**
   * @generated from field: yiru.runtime.v1.GitChangeStatus status = 2;
   */
  status: GitChangeStatus;

  /**
   * @generated from field: optional string old_path = 3;
   */
  oldPath?: string | undefined;

  /**
   * @generated from field: optional uint64 added = 4;
   */
  added?: bigint | undefined;

  /**
   * @generated from field: optional uint64 removed = 5;
   */
  removed?: bigint | undefined;
};
// ...
export enum GitChangeStatus {
  /**
   * @generated from enum value: GIT_CHANGE_STATUS_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: GIT_CHANGE_STATUS_ADDED = 1;
   */
  ADDED = 1,

  /**
   * @generated from enum value: GIT_CHANGE_STATUS_DELETED = 2;
   */
  DELETED = 2,

  /**
   * @generated from enum value: GIT_CHANGE_STATUS_RENAMED = 3;
   */
  RENAMED = 3,

  /**
   * @generated from enum value: GIT_CHANGE_STATUS_COPIED = 4;
   */
  COPIED = 4,

  /**
   * @generated from enum value: GIT_CHANGE_STATUS_MODIFIED = 5;
   */
  MODIFIED = 5,

  /**
   * @generated from enum value: GIT_CHANGE_STATUS_UNTRACKED = 6;
   */
  UNTRACKED = 6,
}